make build.rs be rebuilt upon generator change
This change make build.rs slightly more robust, in particular it informs cargo
that it needs to be rebuilt if the generator python script changes and/or when
the generated file themselves disappear.

It is still not ideal, because currently we are not following the build.rs
recommendation [1] (e.g., we should generate code under env::var_os("OUT_DIR")
and include() it from real Rust modules like in [2], rather than generating the
actual modules under src), but it's an improvement over the status quo.

[1]: https://doc.rust-lang.org/cargo/reference/build-scripts.html
[2]: https://doc.rust-lang.org/cargo/reference/build-script-examples.html