Replace anyhow::Error with enums in bvgraph::load
This allows crates using epserde to unwrap errors and `match` them.

In my case, I would like to display a custom message on WrongTypeHash/VersionMismatch with
instructions to regenerate files with the current epserde version.
Currently this is not enough for this use case (because epserde still
returns anyhow::Error), but a similar change will go to epserde.